move location of utils
